requirements.txt
pillow
pystray
baidu_aip
Pyaudio
wave
keyboard